+++ src/router.ts
@@ -7,6 +7,7 @@ import Developer from './views/Developer.vue';
 import Projects from './views/Projects.vue';
 import Project from './views/Project.vue';
 import Category from './views/Category.vue';
+import Requirement from './views/Requirement.vue';
 import OidcCallback from './views/oidc/OidcCallback.vue';
 import OidcPopupCallback from './views/oidc/OidcPopupCallback.vue'
 import OidcCallbackError from './views/oidc/OidcCallbackError.vue'
@@ -69,7 +70,7 @@ export const router = createRouter({
       component: Projects,
     },
     {
-      path: "/projects/:projectId",
+      path: "/projects/:projectId/:members?",
       name: "project",
       component: Project,
     },
@@ -78,18 +79,35 @@ export const router = createRouter({
       name: "project-all",
       component: Project,
     },
-    {
+/*    {
       path: "/projects/:projectId/members",
       name: "project-members",
       component: Project,
     },
+*/
     {
       path: "/projects/:projectId/categories/:categoryId/:done?",
       name: "category",
       component: Category,
     },
+    {
+      path: "/projects/:projectId/requirements/:requirementId",
+      name: "requirement",
+      component: Requirement,
+    },
   ],
 });
+
+export const routePathToProject = (projectId: string | number) => {
+  return `/projects/${projectId}`;
+};
+
+/**
+ * Returns the *route path* to the given requirment (no absolute URL!).
+ */
+export const routePathToRequirement = (projectId: string | number, requirementId: string | number) => {
+  return `/projects/${projectId}/requirements/${requirementId}`;
+};
 
 router.beforeEach(vuexOidcCreateRouterMiddleware(store, 'oidcStore'))
 router.beforeEach((to, from, next) => {