+++ src/store/state.ts
@@ -1,4 +1,4 @@
-import { Project, Category, Requirement, Comment, Dashboard } from '../types/bazaar-api';
+import { Project, Category, Requirement, Comment, Dashboard, ProjectMember } from '../types/bazaar-api';
 import { Activity } from '../types/activities-api';
 
 export interface LocalComment extends Comment {
@@ -6,17 +6,29 @@ export interface LocalComment extends Comment {
   draftMessage?: String;
 }
 
+export interface UnhandledError {
+  message: string;
+  source: string;
+  details: string;
+  timestamp: number;
+}
+
 export type State = {
   projects: {[id: number]: Project};
+  projectMembers: {
+    [projectId: number]: {[memberId: number]: ProjectMember}
+  };
   categories: {[id: number]: Category};
   requirements: {[id: number]: Requirement};
   comments: {[id: number]: LocalComment};
   activities: {[id: number]: Activity};
   dashboard: Dashboard;
+  unhandledErrors: UnhandledError[];
 }
 
 export const state: State = {
   projects: {},
+  projectMembers: {},
   categories: {},
   requirements: {},
   comments: {},
@@ -26,4 +38,5 @@ export const state: State = {
     categories: [],
     requirements: [],
   },
+  unhandledErrors: [],
 };